% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekbohm.R
\name{ekbohm}
\alias{ekbohm}
\title{Ekbohm's MLE based test under homoscedasticity}
\usage{
ekbohm(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a(non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided"(default), "greater", or"less".}
}
\value{
\item{Z_E}{the value of MLE based Z-statistic.}
\item{p_value}{the p-value for the test.}
}
\description{
Performs a two sample test using the Student's t distribution and MLE statistic under homoscedasticity.
}
\examples{
x <- c(150,210,NA,350,315,NA,NA,98,170,165)
y <- c(NA,150,305,180,NA,90,80,206,111,43)
ekbohm(x,y,alternative="two.sided") #Z_E=1.355812 p_value=0.2331778
ekbohm(x,y,alternative="greater") #Z_E=1.355812  p_value=0.1165889
ekbohm(x,y,alternative="less") #Z_E=1.355812 p_value=0.8834111
}
