% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_stivers.R
\name{lin_stivers}
\alias{lin_stivers}
\title{Lin and Stivers's MLE based test under heteroscedasticity}
\usage{
lin_stivers(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a(non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided"(default), "greater", or"less".}
}
\value{
\item{Z_LS}{the value of MLE based Z-statistic.}
\item{p_value}{the p-value for the test.}
}
\description{
Performs a two sample test using the Student's t distribution and MLE statistic under heteroscedasticity.
}
\examples{
x <- c(150,210,NA,350,315,NA,NA,98,170,165)
y <- c(NA,150,305,180,NA,90,80,206,111,43)
lin_stivers(x,y,alternative="two.sided") #Z_LS=1.584755 p_value=0.1738763
lin_stivers(x,y,alternative="greater") #Z_LS=1.584755  p_value=0.08693813
lin_stivers(x,y,alternative="less") #Z_LS=1.584755 p_value=0.9130619
}
