% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liptak_weighted.R
\name{liptak_weighted}
\alias{liptak_weighted}
\title{Liptak's Weighted Z-test}
\usage{
liptak_weighted(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided", "greater"(default), or "less".}
}
\value{
\item{Z1}{the value of the Z1 test statistic.}
\item{Z2}{the value of the Z2 test statistic.}
\item{p_value}{the p-value for the test.}
}
\description{
Performs a two sample weighted Z-test using a pooled p-value.
}
\examples{
x <- c(150,210,NA,350,315,NA,NA,98,170,165)
y <- c(NA,150,305,180,NA,90,80,206,111,43)
liptak_weighted(x,y,alternative="two.sided") #Z1=0.6227665 Z2=-0.150289 p_value=0.6732314
liptak_weighted(x,y,alternative="greater") #Z1=1.11065 Z2=0.5832401 p_value=0.1068279
liptak_weighted(x,y,alternative="less") #Z1=-1.11065 Z2=-0.5832401 p_value=0.8931721
}
