% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looney_jones.R
\name{looney_jones}
\alias{looney_jones}
\title{Looney and Jones's corrected Z-test}
\usage{
looney_jones(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a(non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided"(default), "greater", or"less".}
}
\value{
\item{Z_corr}{the value of the corrected Z-test statistic.}
\item{p_value}{the p-value for the test.}
}
\description{
Performs a two sample Z-test based on a modified variance estimation of the standard Z-test.
}
\examples{
x <- c(150,210,NA,350,315,NA,NA,98,170,165)
y <- c(NA,150,305,180,NA,90,80,206,111,43)
looney_jones(x,y,alternative="two.sided") #Z_stat=1.430195 p_value=0.152661
looney_jones(x,y,alternative="greater") #Z_stat=1.430195  p_value=0.07633049
looney_jones(x,y,alternative="less") #Z_stat=1.430195 p_value=0.9236695
}
